Fixing the diffusion term.
The balances are right for some reason but term to term comparison fails (and so does the energy).

The problem was that using fvc::div(fvc::grad()) results in 2 interpolation center->face interpolations which is wrong.
It is partially fixed form the mean momentum term. Now fixing it for all the other terms (momentum and energy).